chore: remove old SSL plugin, lrucache and add missing modules to rockspec
